+++ src/components/blog/disqus.module.scss
@@ -1,0 +1,5 @@
+@use '../../sass/variables' as *;
+
+.commentContainer {
+    margin: 50px 35px;
+}
